Restructure project
This merges Library and Core and makes Core independent from Midi so we
can have other export formats too. Also splits common Generator
functionality into a seperate assembly so we can have different types of
generators.

+++ NewWave.Generator/Grooves/FillGenerator.cs
@@ -15,7 +15,7 @@ namespace NewWave.Generator.Grooves
 			var hits = new List<PercussionNote>();
 			foreach (var c in chain)
 			{
-				var theseHits = Enumerable.Range(0, (int)(c.Length / lengthOfNote)).Select(n => new PercussionNote(delay + n * lengthOfNote, c.Note, n == 0 ? Velocity.Fff : Velocity.F)).ToList();
+				var theseHits = Enumerable.Range(0, (int)(c.Length / lengthOfNote)).Select(n => new PercussionNote(delay + n * lengthOfNote, (int) c.Note, (int) (n == 0 ? Velocity.Fff : Velocity.F))).ToList();
 				hits.AddRange(theseHits);
 				delay += c.Length;
 			}
